Fixed Create and Delete operation of Question and TestCases Model

+++ src/questions/questions.service.ts
@@ -1,5 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { TestCase } from 'src/test-cases/entities';
 import { Repository } from 'typeorm';
 import { CreateQuestionDTO } from './dto/create-question.dto';
 import { UpdateQuestionDTO } from './dto/update-question.dto';
@@ -13,15 +14,26 @@ export class QuestionsService {
   ) {}
 
   findAll(): Promise<Question[]> {
-    return this.questionRepository.find();
+    return this.questionRepository.find({
+      relations: {
+        test_cases: true,
+      },
+    });
   }
 
   findOne(id: number): Promise<Question> {
     return this.questionRepository.findOneBy({ id });
   }
 
-  async create(newQuestion: CreateQuestionDTO): Promise<Question> {
-    return await this.questionRepository.save(newQuestion);
+  async create(newQuestion: CreateQuestionDTO) {
+    newQuestion.test_cases = newQuestion.test_cases.map((testCase) => {
+      const tCase = new TestCase();
+      tCase.stdin = testCase.stdin;
+      tCase.output = testCase.output;
+      return tCase;
+    });
+    console.log(newQuestion);
+    return this.questionRepository.save(newQuestion);
   }
 
   async update(id: number, updQuestion: UpdateQuestionDTO): Promise<Question> {
@@ -36,8 +48,12 @@ export class QuestionsService {
     return await this.questionRepository.save(question);
   }
 
-  async remove(id: number): Promise<Question> {
+  async remove(id: number) {
     const question = await this.findOne(id);
+    console.log(question);
+    if (!question) {
+      throw new NotFoundException();
+    }
     return this.questionRepository.remove(question);
   }
 }
